feat: move space creation from FlatfileProvider to component
