+++ packages/react/src/hooks/useFlatfile.ts
@@ -1,5 +1,7 @@
 import { useContext } from 'react'
 import FlatfileContext from '../components/FlatfileContext'
+import { initializeSpace } from '../utils/initializeSpace'
+import { getSpace } from '../utils/getSpace'
 
 export const useFlatfile = () => {
   const context = useContext(FlatfileContext)
@@ -9,7 +11,50 @@ export const useFlatfile = () => {
     throw new Error('useFlatfile must be used within a FlatfileProvider')
   }
 
+  const {
+    publishableKey,
+    environmentId,
+    apiUrl,
+    space,
+    setSessionSpace,
+    accessToken,
+    setAccessToken,
+  } = context
+
+  const createSpace = async () => {
+    if (publishableKey) {
+      const createdSpace = await initializeSpace({
+        publishableKey,
+        environmentId,
+        apiUrl,
+      })
+      if (createdSpace?.data.accessToken) {
+        ;(window as any).CROSSENV_FLATFILE_API_KEY =
+          createdSpace?.data.accessToken
+        setAccessToken(createdSpace?.data.accessToken)
+        setSessionSpace(createdSpace)
+      }
+    }
+  }
+
+  const reUseSpace = async () => {
+    if (space) {
+      const reUsedSpace = await getSpace({
+        space,
+        environmentId,
+        apiUrl,
+      })
+      setAccessToken(space.accessToken)
+      setSessionSpace(reUsedSpace)
+    }
+  }
+
   const openPortal = () => {
+    if (!!publishableKey) {
+      createSpace()
+    } else if (space) {
+      reUseSpace()
+    }
     setOpen(true)
   }
