+++ packages/react/src/utils/addSpaceInfo.ts
@@ -18,7 +18,6 @@ export const addSpaceInfo = async (
 
   try {
     if (!workbook && sheet) {
-      console.log({ workbook, sheet })
       const createdWorkbook = createWorkbookFromSheet(sheet, !!onSubmit)
       localWorkbook = await api.workbooks.create({
         spaceId,
